refactor: hoist safeReturnPath to @kuruma/shared + harden OAuth callback (#510)
Code-review follow-ups:
- Move the open-redirect guard into @kuruma/shared/lib/return-path so the API
  and web boundary share ONE definition (was duplicated -> drift risk on a
  security primitive). Re-exported from auth/google.ts; web imports from shared.
- Origin-qualify the callback redirect: resolve the validated root-relative
  returnTo against postLoginRedirect's web origin so it can't land on the API
  origin off-proxy (#378 cutover insurance).
- Clear both OAuth-flow cookies on the callback 400 paths (no lingering state).
