+++ packages/api/src/auth/google.ts
@@ -57,31 +57,10 @@ export const OAUTH_STATE_TTL_SECONDS = 600
  *  survives Google's top-level redirect and expires with the flow. */
 export const OAUTH_RETURN_COOKIE = 'kuruma_oauth_return'
 
-const MAX_RETURN_PATH_LENGTH = 512
-const CONTROL_CHAR_MAX = 0x1f
-const DEL_CHAR = 0x7f
-
-/**
- * Validate a post-login redirect target as a *local* path — the open-redirect
- * defence. Accepts only same-origin root-relative paths (`/en/bookings`);
- * rejects protocol-relative (`//evil`), absolute URLs, backslash tricks
- * (browsers fold `\` to `/`), control chars (CR/LF header injection), and
- * anything not starting with a single `/`. Returns the trusted path, or
- * undefined to fall back to the configured postLoginRedirect.
- */
-export function safeReturnPath(raw: string | null | undefined): string | undefined {
-  if (typeof raw !== 'string' || raw.length === 0 || raw.length > MAX_RETURN_PATH_LENGTH) {
-    return undefined
-  }
-  if (raw[0] !== '/') return undefined
-  if (raw[1] === '/' || raw[1] === '\\') return undefined
-  if (raw.includes('\\')) return undefined
-  for (const ch of raw) {
-    const code = ch.charCodeAt(0)
-    if (code <= CONTROL_CHAR_MAX || code === DEL_CHAR) return undefined
-  }
-  return raw
-}
+// The open-redirect guard for `returnTo` lives in @kuruma/shared so the API and
+// the web boundary enforce ONE definition — a guard cloned across a trust
+// boundary drifts. Re-exported so the auth routes keep importing it from here.
+export { safeReturnPath } from '@kuruma/shared/lib/return-path'
 
 export interface GoogleOAuthConfig {
   readonly clientId: string
